refactor the code completly
