+++ src/extractors/base_extractor.py
