+++ src/transformer/categorizer.py
@@ -1,0 +1,2 @@
+class Categorizer:
+    pass
