+++ src/transformer/deduplicator.py
@@ -1,0 +1,2 @@
+class Deduplicator:
+    pass
